driveparts/requirements.txt
analysis==0.1.7
asgiref==3.8.1
asttokens==2.4.1
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
build==1.2.2.post1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.2
django-allauth==65.4.1
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.1
idna==3.10
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
oauthlib==3.2.2
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pickleshare==0.7.5
pillow==11.1.0
pip-tools==7.4.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pyzmq==26.2.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
setuptools==75.8.2
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
transliterate==1.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
yarg==0.1.9

gevent==24.11.1
gunicorn==23.0.0
